[EPMEDU-864]: Extract screens of more flow to separate modules
