+++ feature/mainflow/moreflow/about/build.gradle.kts
@@ -4,22 +4,12 @@ plugins {
 }
 
 dependencies {
-    implementation(projects.feature.mainflow.moreflow.flow)
-    implementation(projects.feature.home)
-
     implementation(projects.library.extensions)
     implementation(projects.library.foundation)
     implementation(projects.library.navigation)
     implementation(projects.library.resources)
 
-    implementation(libs.accompanist.permissions)
-
-    implementation(libs.androidx.viewmodel)
-    implementation(libs.androidx.viewmodel.compose)
-
     implementation(libs.compose.ui)
     implementation(libs.compose.ui.preview)
-    implementation(libs.compose.material)
-
     debugImplementation(libs.compose.ui.tooling)
 }
