+++ feature/mainflow/moreflow/about/src/main/java/com/epmedu/animeal/more/about/AboutScreen.kt
@@ -1,0 +1,18 @@
+package com.epmedu.animeal.more.about
+
+import androidx.compose.runtime.Composable
+import com.epmedu.animeal.extensions.currentOrThrow
+import com.epmedu.animeal.foundation.bottombar.BottomBarVisibility
+import com.epmedu.animeal.foundation.bottombar.BottomBarVisibilityState.HIDDEN
+import com.epmedu.animeal.navigation.navigator.LocalNavigator
+
+@Composable
+fun AboutScreen() {
+    val navigator = LocalNavigator.currentOrThrow
+
+    BottomBarVisibility(HIDDEN)
+
+    AboutScreenUI(
+        onBack = navigator::popBackStack,
+    )
+}
