+++ feature/mainflow/moreflow/donate/src/main/java/com/epmedu/animeal/more/donate/DonateScreen.kt
@@ -1,0 +1,18 @@
+package com.epmedu.animeal.more.donate
+
+import androidx.compose.runtime.Composable
+import com.epmedu.animeal.extensions.currentOrThrow
+import com.epmedu.animeal.foundation.bottombar.BottomBarVisibility
+import com.epmedu.animeal.foundation.bottombar.BottomBarVisibilityState.HIDDEN
+import com.epmedu.animeal.navigation.navigator.LocalNavigator
+
+@Composable
+fun DonateScreen() {
+    val navigator = LocalNavigator.currentOrThrow
+
+    BottomBarVisibility(HIDDEN)
+
+    DonateScreenUI(
+        onBack = navigator::popBackStack,
+    )
+}
