+++ feature/mainflow/moreflow/help/build.gradle.kts
@@ -1,25 +1,21 @@
 plugins {
     id("AnimealPlugin")
     id("com.android.library")
+    id("kotlin-kapt")
+    id("dagger.hilt.android.plugin")
 }
 
 dependencies {
-    implementation(projects.feature.mainflow.moreflow.flow)
-    implementation(projects.feature.home)
-
+    implementation(projects.library.common)
     implementation(projects.library.extensions)
     implementation(projects.library.foundation)
     implementation(projects.library.navigation)
     implementation(projects.library.resources)
 
-    implementation(libs.accompanist.permissions)
-
-    implementation(libs.androidx.viewmodel)
-    implementation(libs.androidx.viewmodel.compose)
-
+    implementation(libs.compose.material)
     implementation(libs.compose.ui)
     implementation(libs.compose.ui.preview)
-    implementation(libs.compose.material)
-
     debugImplementation(libs.compose.ui.tooling)
+    implementation(libs.hilt.android)
+    kapt(libs.hilt.compiler)
 }
