+++ feature/mainflow/moreflow/help/src/main/java/com/epmedu/animeal/more/help/HelpScreen.kt
@@ -1,25 +1,26 @@
-package com.epmedu.animeal.more.profile.presentation
+package com.epmedu.animeal.more.help
 
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
 import androidx.hilt.navigation.compose.hiltViewModel
 import com.epmedu.animeal.extensions.currentOrThrow
 import com.epmedu.animeal.foundation.bottombar.BottomBarVisibility
 import com.epmedu.animeal.foundation.bottombar.BottomBarVisibilityState.HIDDEN
-import com.epmedu.animeal.more.profile.presentation.viewmodel.ProfileViewModel
+import com.epmedu.animeal.more.help.viewmodel.HelpViewModel
 import com.epmedu.animeal.navigation.navigator.LocalNavigator
 
 @Composable
-fun ProfileScreen() {
-    val viewModel: ProfileViewModel = hiltViewModel()
+fun HelpScreen() {
     val navigator = LocalNavigator.currentOrThrow
-    val state = viewModel.stateFlow.collectAsState().value
+    val helpViewModel = hiltViewModel<HelpViewModel>()
+
+    val state by helpViewModel.stateFlow.collectAsState()
 
     BottomBarVisibility(HIDDEN)
 
-    ProfileScreenUI(
-        state = state,
+    HelpScreenUI(
+        helpState = state,
         onBack = navigator::popBackStack,
-        onEvent = viewModel::handleEvent
     )
 }
