+++ feature/mainflow/moreflow/help/src/main/java/com/epmedu/animeal/more/help/viewmodel/HelpState.kt
@@ -1,0 +1,5 @@
+package com.epmedu.animeal.more.help.viewmodel
+
+data class HelpState(
+    val appVersionName: String = ""
+)
