--- feature/more/src/main/java/com/epmedu/animeal/more/help/HelpState.kt
@@ -1,5 +1,0 @@
-package com.epmedu.animeal.more.help
-
-data class HelpState(
-    val appVersionName: String = ""
-)
