+++ library/navigation/src/main/java/com/epmedu/animeal/navigation/route/MoreRoute.kt
@@ -1,0 +1,10 @@
+package com.epmedu.animeal.navigation.route
+
+enum class MoreRoute {
+    More,
+    Profile,
+    Donate,
+    Help,
+    About,
+    Account
+}
