+++ settings.gradle.kts
@@ -44,6 +44,13 @@ rootProject.name = "Animeal"
 
 include(":app")
 
+include(":feature:home")
+include(":feature:mainflow:moreflow:about")
+include(":feature:mainflow:moreflow:account")
+include(":feature:mainflow:moreflow:flow")
+include(":feature:mainflow:moreflow:more")
+include(":feature:mainflow:moreflow:help")
+include(":feature:mainflow:moreflow:donate")
 include(":feature:more")
 include(":feature:signupflow:entercode")
 include(":feature:signupflow:enterphone")
@@ -52,7 +59,6 @@ include(":feature:signupflow:onboarding")
 include(":feature:signupflow:signup")
 include(":feature:splash")
 include(":feature:tabs")
-include(":feature:home")
 
 include(":library:common")
 include(":library:extensions")
